+++ tests/tipc.nim
@@ -1,0 +1,2 @@
+when defined(feature.sigils.ipc):
+  include ipcCases